fix(refactor): move a few things around ... larger refact coming
