+++ packages/vehicles/src/index.js
@@ -4,59 +4,12 @@ import React from "react";
 import PropTypes from "prop-types";
 
 import VehicleLayer from "./VehicleLayer";
-import { checkRefreshInteval } from "./utils";
+import { checkRefreshInteval, fetchVehicles } from "./utils";
 
-function getVehicles(setVehicleData, setTrackedVehicle, trackId, url) {
-  const d = Date.now();
-  url = url || "https://maps.trimet.org/gtfs/rt/vehicles/routes/all";
-  url = url.indexOf("?") ? `${url}?` : `${url}&`;
-  url = `${url}__time__=${d}`;
-
-  fetch(url)
-    .then(res => {
-      if (!res.ok) {
-        console.log(res.statusText);
-        throw Error(res.statusText);
-      }
-      return res;
-    })
-    .then(res => {
-      const retVal = res.json();
-      return retVal;
-    })
-    .then(vehicleList => {
-      if (vehicleList && vehicleList.length > 0) {
-        // step 1: set the vehicle list (triggers vehicle points redraw)
-        console.log(`updating state with ${vehicleList.length} vehicles`);
-        setVehicleData(vehicleList);
-
-        if (trackId) {
-          let tracked = null;
-
-          // step 2: find vehicle record via either tripId or vehicleId
-          vehicleList.some(v => {
-            if (trackId === v.vehicleId || trackId === v.tripId) {
-              tracked = v;
-              return true;
-            }
-            return false;
-          });
-          // step 3: add updated tracked vehicle to state (triggers pattern line redraw)
-          if (tracked) {
-            setTrackedVehicle(tracked);
-          } else {
-            console.log(`WARN: can't find tripId or vehicleId ${trackId}`);
-          }
-        }
-      } else {
-        console.log("get vehicle data is suspect");
-      }
-    })
-    .catch(error => {
-      console.log(`VEH fetch() error: ${error}`);
-    });
-}
-
+/**
+ * TODO: describe
+ * TODO: talk about gtfsdb and opensource
+ */
 function Vehicles(props) {
   const refreshDelay = checkRefreshInteval(props.refreshDelay);
 
@@ -68,6 +21,26 @@ function Vehicles(props) {
   React.useEffect(() => {
     trackedVehicleRef.current = trackedVehicle;
   }, [trackedVehicle]);
+
+  /** callback function where the setTrackedVehicle is stored */
+  function setTrackedData(vehicle) {
+    // todo check/grab geometry ... maybe delay setTrackedVehicle until we have said geom
+    /*
+    // step 1: vehicle geometry
+    if(vehicle && vehicle.patternId) {
+      // setVehicleGeometry(vehicle);
+      const [trackedVehicleGeometry, setTrackedVehicleGeometry] = React.useState(null);
+      if(cache[patternId] === null) {
+          getVehicleGeometry(setTrackedVehicleGeometry, patternId);
+      } else {
+          setTrackedVehicleGeometry(cache[patternId]);
+      }
+    */
+    // TODO: should setTrackedVehicle be both a vehicle and the route geometry?
+    // setTrackedData({v: vehicle, g: geometry})?  ... thus only one state / one useEffect/Ref
+    // step 2: set tracked vehicle
+    setTrackedVehicle(vehicle);
+  }
 
   function getTrackedVehicle() {
     let retVal = null;
@@ -91,9 +64,9 @@ function Vehicles(props) {
     //       if we don't have the gate of vehicleData == null, then we'll get multiple setInterval calls
     let interval = null;
     if (vehicleData == null) {
-      getVehicles(setVehicleData, setTrackedVehicle, props.tracked);
+      fetchVehicles(setVehicleData, setTrackedVehicle, props.tracked);
       interval = setInterval(() => {
-        getVehicles(setVehicleData, setTrackedVehicle, getTrackedVehicleId());
+        fetchVehicles(setVehicleData, setTrackedData, getTrackedVehicleId());
       }, refreshDelay);
     }
